Fixed some minor build warnings with format specifiers and data types in the Benchmarks code.
git-svn-id: http://dysart.cs.byu.edu/chsvn/trunk@540 a75b6edf-6070-4da8-89c8-9992cc802e51

+++ test/Benchmarks.m
@@ -651,9 +651,9 @@ int main (int argc, const char * argv[]) {
 	NSUInteger scale = 1000000; // 10^6, which gives microseconds
 	
 	for (NSUInteger trial = 1; trial <= reps; trial++) {
-		printf("\nPass %lu / %lu", trial, reps);
+		printf("\nPass %lu / %lu", (unsigned long)trial, (unsigned long)reps);
 		for (NSUInteger size = 10; size <= limit; size *= 10) {
-			printf("\n%8lu objects --", size);
+			printf("\n%8lu objects --", (unsigned long)size);
 			// Create a set of N unique random numbers
 			NSAutoreleasePool *pool = [[NSAutoreleasePool alloc] init];
 			while ([objectSet count] < size)
